I worked on adding a call to set chat.channel to null when current channel is deleted, right before default channel is selected. This was done in anticipation of the possible event that this is the last channel that was in the org. This required the following updates: - Now only displaying chat component if type channel with existing channel, or type private with existing partnerUsername (added this to protect against errors when last channel was deleted) - Added chat.setChannel() action and SET_CHANNEL type (called by sidebar action selectChannel). I added this so that the channel could be set to null when the current channel was deleted, right before selecting default channel (if applicable - if any more channels even exist) - Only fetching channel member permissions if we have a channel selected (not null)
The other thing I worked on was a bug fix. When channel is deleted to find out if it's the current channel, it's not enough to check if the name matches up with chat.channel. We now check if the org names match as well

On switching orgs we now reset chat before selecting the org. The new org may not have any channels and we dont want chat populated with old info from previous org's chat